PhotoGallery: Use vehicle/flight metadata
Extract vehicle and flight metadata from images pushed into the gallery.
Store these images in folder structure (vehicle/flight/image). Make sure
that storage layers deals with on-disk structure containing a mixture
of images with/without metadata tags.

Expose image metadata as part of PhotoGalleryModel API.

+++ src/PhotoGallery/PhotoFileStore.cc
@@ -14,6 +14,8 @@
 #include <QDebug>
 #include <QtQml>
 
+#include "ExtractJPEGMetadata.h"
+
 PhotoFileStore::PhotoFileStore(QObject * parent)
     : QObject(parent)
 {
@@ -62,11 +64,23 @@ QString PhotoFileStore::add(QString name_hint, QByteArray data)
     if (_location.isEmpty()) {
         return {};
     }
-    QString id = name_hint;
+
+    auto metadata = extractJPEGMetadata(data);
+    QString base = "";
+    auto maybe_flight_uid = metadata.find("PX4:flight_uid");
+    auto maybe_vehicle_uuid = metadata.find("PX4:vehicle_uuid");
+    if (maybe_flight_uid != metadata.end() && maybe_vehicle_uuid != metadata.end()) {
+        base = maybe_vehicle_uuid->second + QDir::separator() + maybe_flight_uid->second;
+        QDir(_location).mkpath(base);
+        base = base + QDir::separator();
+    }
+
+    QString id = base + name_hint;
+
     while (_photo_ids.find(id) != _photo_ids.end()) {
         QDateTime now = QDateTime::currentDateTime();
         QString suffix = QFileInfo(name_hint).suffix();
-        id = now.toString("yyyy-MM-dd-HH-mm-ss.z");
+        id = base + now.toString("yyyy-MM-dd-HH-mm-ss.z");
         if (!suffix.isEmpty()) {
             id += "." + suffix;
         }
@@ -92,10 +106,27 @@ void PhotoFileStore::remove(const std::set<QString> & ids)
     if (_location.isEmpty()) {
         return;
     }
+
+    QDir base(_location);
+
     for (const auto & id : ids) {
-        _photo_ids.erase(id);
-        QFile file(QDir(_location).filePath(id));
+        std::size_t count = _photo_ids.erase(id);
+        if (count == 0) {
+            // Tried to erase an image that does not exist. Let's not pass this
+            // operation down to the filesystem.
+            continue;
+        }
+        QFile file(base.filePath(id));
         file.remove();
+
+        // Try to remove directories that this file was in.
+        QString path = id;
+        int sep = path.lastIndexOf('/');
+        if (sep == -1) {
+            break;
+        }
+        path.remove(sep, path.size() - sep);
+        base.rmpath(path);
     }
     emit removed(ids);
 }
@@ -118,12 +149,13 @@ void PhotoFileStore::rescan()
 {
     emit removed(_photo_ids);
     _photo_ids.clear();
-    auto entries = QDir(_location).entryInfoList();
-    for (const auto & entry : entries) {
-        if (!entry.isFile()) {
-            continue;
-        }
-        _photo_ids.insert(entry.fileName());
+    QDir base(_location);
+    QDirIterator iterator(_location, QStringList("*"),  QDir::Files, QDirIterator::Subdirectories);
+    while (iterator.hasNext()) {
+        iterator.next();
+        QFileInfo info = iterator.fileInfo();
+        QString relpath = base.relativeFilePath(info.absoluteFilePath());
+        _photo_ids.insert(relpath);
     }
     emit added(_photo_ids);
 }
